requirements.txt
argcomplete==3.1.1
asgiref==3.8.1
click==8.1.7
Django==5.1.5
django-taggit==6.1.0
pipx==1.2.0
sqlparse==0.5.3
userpath==1.9.0
